backend/requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
tensorflow==2.8.1
python-multipart==0.0.5
pillow==9.5.0
#for template  #new
jinja2==3.0.3
aiofiles
protobuf==3.20.*